% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFields.R
\docType{data}
\name{dataFields}
\alias{dataFields}
\title{Fields in the Vermont Atlas of Life Solr database.}
\format{
An object of class \code{data.frame} with 237 rows and 14 columns.
}
\usage{
dataFields
}
\value{
A data frame with 237 rows and 14 variables:
\describe{
  \item{name}{Variable name}
  \item{dataType}{How are data stored}
  \item{indexed}{is the data indexed in the db}
  \item{stored}{are the data stored or computed}
  \item{multivalue}{does the field return a multivalue field}
  \item{docvalue}{??}
  \item{downloadName}{Name needed to get data from API}
  \item{description}{simple description of the field}
  \item{dwcTerm}{Darwin Core name}
  \item{classs}{??}
  \item{jsonName}{??}
  \item{downloadDescription}{??}
  \item{i18nValues}{??}
  \item{info}{description and/or link to more information}
}
}
\description{
A dataset containing the field names and attributes of the
solr database as of April 2021.
}
\keyword{datasets}
