% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSON_to_dataframe.R
\name{JSON_to_dataframe}
\alias{JSON_to_dataframe}
\title{Collate content data from API request}
\usage{
JSON_to_dataframe(x, fields = "minimal", type = "occurrence", quiet = TRUE)
}
\arguments{
\item{x}{is the content from an API query}

\item{fields}{which data fields should be returned? \code{all} returns all
fields that are not \code{class("list")}, while \code{minimal} returns only
a subset of fields to run analyses - most metadata for the observation are
removed. However, enough information is retained to relate the obs. back to
the appropriate metadata.}

\item{type}{which records to report}

\item{quiet}{defaults to \code{TRUE}, if \code{FALSE} prints the number of
records and number of pages to console}
}
\value{
\code{data.frame} where each observation is a row
}
\description{
\code{JSON_to_dataframe} extracts data from an API request and reformats
into a dataframe
}
