% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_observations.R
\name{map_observations}
\alias{map_observations}
\title{Generate heatmap for the number of observations within polygons}
\usage{
map_observations(
  query,
  polygons,
  number_obs = TRUE,
  seq_by = NULL,
  cols = NULL,
  plot_results = FALSE
)
}
\arguments{
\item{query}{\code{list} of query terms \code{list('order:Odonata','etc')}}

\item{polygons}{shapefile of \code{POLYGON} or \code{MULTIPOLYGON}, can be
either an \code{sf} object or \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}}

\item{seq_by}{an integer used to 'group' the number of observations. Defaults to 10}

\item{cols}{should be a vector of colors used for the color ramp}

\item{plot_results}{boolean. When \code{TRUE} a simple plot is returned.}
}
\value{
polygon with NumObs and when \code{plot_results=TRUE} a map of the number of observations
}
\description{
Generate heatmap for the number of observations within polygons
}
\examples{
\dontrun{
SurveyBlocks <- sf::st_read("VT_survey_block_WGS.shp", quiet = TRUE)
z <- map_observations(query = list("order:Odonata"),
                          polygons = SurveyBlocks,
                          seq_by = 10,
                          cols = c("gray","blue","firebrick"))}
}
\seealso{
\code{queryAPI} for description of fields and which fields can be queried
}
