% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{queryAPI}
\alias{queryAPI}
\title{Create query to grab data from API request}
\usage{
queryAPI(
  query,
  type = "occurrences",
  fields = "all",
  quiet = FALSE,
  pageSize = NULL,
  spatialWKT = NULL
)
}
\arguments{
\item{query}{\code{list} of query terms \code{list('order:Odonata','etc')}}

\item{type}{options include "TotalObs" and 'occurrences'. TotalObs returns only the number of observations. No data.
occurrence returns data associated with occurrence records. default = occurrences.}

\item{fields}{defaults to 'all'. see \code{dataFields} for returnable fields}

\item{quiet}{defaults to \code{FALSE}, if \code{TRUE} prints brief description to console}

\item{pageSize}{defaults to NULL which returns all records}

\item{spatialWKT}{well-known text for a \code{POLYGON} or \code{MULTIPOLYGON}, defaults to null}
}
\value{
\code{data frame} of results from query
}
\description{
\code{get_data} is a wrapper for httr::GET that connects to Vermont Atlas of
Life's API
}
\examples{
\dontrun{queryAPI(query = list("order:Odonata"),
                  type = 'occurrences',
                  fields = c('taxon_name','longitude','latitude','year'))}
}
\seealso{
\code{dataFields} for description of fields and which fields can be queried
}
