% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towns.R
\docType{data}
\name{towns}
\alias{towns}
\title{Vermont Towns.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 255 rows and 22 columns.
}
\usage{
towns
}
\value{
Simple feature collection with 255 features and 21 fields
\describe{
  \item{AREA}{Area of polygon in m}
  \item{PERIMETER}{Perimeter of polygon in km}
  \item{TOWNS_}{numeric}
  \item{TOWNS_ID}{numeric}
  \item{FIPS6}{FIPS code - 6digits}
  \item{MCODE}{Mail code}
  \item{TOWNNAME}{Name of the town}
  \item{PSAP}{??}
  \item{MDPATH}{??}
  \item{CERT}{??}
  \item{CERTDT}{??}
  \item{RPC}{??}
  \item{ODDSIDE}{??}
  \item{INCTYPE}{description and/or link to more information}
  \item{...}{...}
  \item{geometry}{spatial type and spatial information}
}
}
\description{
A spatial layer with the towns of Vermont.
}
\keyword{datasets}
