% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFields.R
\docType{data}
\name{dataFields}
\alias{dataFields}
\title{Fields in the Vermont Atlas of Life Solr database.}
\format{
An object of class \code{data.frame} with 237 rows and 14 columns.
}
\usage{
dataFields
}
\value{

}
\description{
A dataset containing the field names and attributes of the
solr database as of April 2021.
}
\keyword{datasets}
