% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{queryAPI}
\alias{queryAPI}
\title{Create query to grab data from API request}
\usage{
queryAPI(
  query,
  type = "occurrences",
  fields = "all",
  quiet = FALSE,
  pageSize = NULL,
  spatialWKT = NULL
)
}
\arguments{
\item{query}{\code{list} of query terms \code{list('order:Odonata','etc')}}

\item{type}{default = occurrences}

\item{fields}{defaults to 'all'}

\item{quiet}{defaults to \code{FALSE}, if \code{TRUE} prints brief description to console}

\item{pageSize}{defaults to NULL which returns all records}

\item{spatialWKT}{well-known text for a \code{POLYGON} or \code{MULTIPOLYGON}, defaults to null}
}
\value{
\code{data frame} of results from query
}
\description{
\code{get_data} is a wrapper for httr::GET that connects to Vermont Atlas of
Life's API
}
\seealso{
\code{dataFields} for description of fields and which fields can be queried
}
